Latest commits, reworked how chatTest latency works
* updates to the entire screen can happen less often
* generation is still realtime

+++ base/common.h
@@ -67,7 +67,7 @@ struct gpt_params {
     float   yarn_beta_fast      = 32.0f;// YaRN low correction dim
     float   yarn_beta_slow      = 1.0f; // YaRN high correction dim
     int32_t yarn_orig_ctx       = 0;    // YaRN original context length
-    int8_t  rope_scaling_type   = LLAMA_ROPE_SCALING_UNSPECIFIED;
+    int32_t rope_scaling_type   = LLAMA_ROPE_SCALING_UNSPECIFIED;
 
     // // sampling parameters
     struct llama_sampling_params sparams;
